+++ src/pages/profile/index.tsx
@@ -1,9 +1,14 @@
 import { Outlet } from '@tanstack/react-router';
+import { Suspense } from 'react';
+
+import { Loader } from '@shared/ui/loader';
 
 export const Profile = (): React.JSX.Element => {
   return (
-    <div className="flex flex-1">
-      <Outlet />
-    </div>
+    <Suspense fallback={<Loader />}>
+      <div className="flex flex-1">
+        <Outlet />
+      </div>
+    </Suspense>
   );
 };
